requirements.txt
requests
pytz
lightstreamer-client
streamlit
pandas
